tests_generated/test_use_github_actions-True_deployment_language-typescript/pyproject.toml
[tool.poetry]
name = "test_use_github_actions-True_deployment_language-typescript"
version = "0.1.0"
description = "Algorand smart contracts"
authors = ["None <None>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
beaker-pyteal = "^1.0.0"
algokit-utils = {version = "^1.2.0b8", allow-prereleases = true}
python-dotenv = "^1.0.0"

[tool.poetry.group.dev.dependencies]
black = {extras = ["d"], version = "*"}
ruff = ">=0.0.260"
mypy = "*"
pytest = "*"
pytest-cov = "*"
pip-audit = "*"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
select = ["E", "F", "ANN", "UP", "N", "C4", "B", "A", "YTT", "W", "FBT", "Q", "RUF", "I"]
ignore = [
  "ANN101", # no type for self
  "ANN102", # no type for cls
]
unfixable = ["B", "RUF"]

[tool.ruff.flake8-annotations]
allow-star-arg-any = true
suppress-none-returning = true

[tool.pytest.ini_options]
pythonpath = ["smart_contracts", "tests"]

[tool.mypy]
files = "smart_contracts/"
python_version = "3.10"
check_untyped_defs = true
warn_redundant_casts = true
warn_unused_ignores = true
allow_untyped_defs = false
strict_equality = true
